requirements.txt
fastapi
uvicorn
aioredis
python-dotenv
asyncpg
psycopg2